refactor: Remove index in context / Change import statement
