--- src/context/index.js
@@ -1,3 +1,0 @@
-import GameContext, { GameContextProvider } from "./GameContext";
-
-export { GameContext, GameContextProvider };
